+++ Providers/AdobeProvider.m
@@ -2,10 +2,27 @@
 #import "ARAnalyticsProviders.h"
 #import "ADBMobile.h"
 
+@interface AdobeProvider ()
+/**
+ * The default implementation for this provider is to track screens "-trackAction:data:" calls when they contain
+ * custom key/value properties in addition to the pageTitle. By setting this property to YES, the Adobe Provider
+ * will call "-trackState:data:" for page views that contain additional custom properties.
+ *
+ * The default value is NO.
+ */
+@property (nonatomic) BOOL callsTrackStateForPageViewsWithProperties;
+@end
+
+NSString const *AdobeProviderCallsTrackStateForPageViewsWithProperties = @"callsTrackStateForPageViewsWithProperties";
+
 @implementation AdobeProvider
 #ifdef AR_ADOBE_EXISTS
 
 - (id)initWithData:(NSDictionary *)additionalData {
+    return [self initWithData:additionalData settings:nil];
+}
+
+- (id)initWithData:(NSDictionary *)additionalData settings:(NSDictionary*)settings {
     NSAssert([ADBMobile class], @"Adobe is not included");
     if(additionalData) {
         [ADBMobile collectLifecycleDataWithAdditionalData:additionalData];
@@ -13,7 +30,11 @@
         [ADBMobile collectLifecycleData];
     }
 
-    return [super init];
+    self = [super init];
+    if (self) {
+        self.callsTrackStateForPageViewsWithProperties = [settings[AdobeProviderCallsTrackStateForPageViewsWithProperties] boolValue];
+    }
+    return self;
 }
 
 - (id)initWithIdentifier:(NSString *)identifier
@@ -30,6 +51,16 @@
     [ADBMobile trackState:pageTitle
                      data:nil];
 }
+
+- (void)didShowNewPageView:(NSString *)pageTitle withProperties:(NSDictionary *)properties {
+    if (self.callsTrackStateForPageViewsWithProperties) {
+        [ADBMobile trackState:pageTitle
+                        data:properties];
+    } else {
+        [super didShowNewPageView:pageTitle withProperties:properties];
+    }
+}
+
 
 #endif
 @end